Oom fallback script (#1809)
* add recover mode in qdrant + script for handelling OOM

* fix clippy

* reformat entrypoint.sh

+++ tests/low-ram/low-ram.sh
@@ -4,13 +4,36 @@ set -euo pipefail
 
 cd "$(dirname "${BASH_SOURCE[0]}")"
 
-docker build ../../ --tag=qdrant_consensus
+declare DOCKER_IMAGE_NAME=qdrant-recovery
 
-if [[ ! -e storage ]]
-then
+docker build ../../ --tag=$DOCKER_IMAGE_NAME
+
+if [[ ! -e storage ]]; then
     git lfs pull
     tar -xf storage.tar.xz
 fi
+
+declare OOM_CONTAINER_NAME=qdrant-oom
+
+docker rm -f ${OOM_CONTAINER_NAME} || true
+
+docker run \
+    -m 128m \
+    -p 127.0.0.1:6333:6333 \
+    -p 127.0.0.1:6334:6334 \
+    -v $PWD/storage:/qdrant/storage \
+    -e QDRANT__STORAGE__HANDLE_COLLECTION_LOAD_ERRORS=true \
+    --name ${OOM_CONTAINER_NAME} \
+    $DOCKER_IMAGE_NAME || true
+
+declare oom_status_code && oom_status_code=$(docker inspect OOM_CONTAINER_NAME --format='{{.State.ExitCode}}')
+
+if ((${oom_status_code} != "137")); then
+    echo "Expected qdrant to OOM" >&2
+    exit 1
+fi
+
+docker rm -f ${OOM_CONTAINER_NAME}
 
 declare container && container=$(
     docker run --rm -d \
@@ -19,7 +42,8 @@ declare container && container=$(
         -p 127.0.0.1:6334:6334 \
         -v $PWD/storage:/qdrant/storage \
         -e QDRANT__STORAGE__HANDLE_COLLECTION_LOAD_ERRORS=true \
-        qdrant_consensus
+        -e QDRANT_ALLOW_RECOVERY_MODE=true \
+        $DOCKER_IMAGE_NAME
 )
 
 function cleanup {
@@ -30,10 +54,8 @@ trap cleanup EXIT
 
 # Wait (up to ~30 seconds) for the service to start
 declare retry=0
-while [[ $(curl -sS localhost:6333 -w ''%{http_code}'' -o /dev/null) != 200 ]]
-do
-    if (( retry++ < 30 ))
-    then
+while [[ $(curl -sS localhost:6333 -w ''%{http_code}'' -o /dev/null) != 200 ]]; do
+    if ((retry++ < 30)); then
         sleep 1
     else
         echo "Service failed to start in ~30 seconds" >&2
@@ -43,10 +65,8 @@ done
 
 # Wait (up to ~10 seconds) until `low-ram` collection is loaded
 declare retry=0
-while ! curl -sS --fail-with-body localhost:6333/collections | jq -e '.result.collections | index({"name": "low-ram"})' &>/dev/null
-do
-    if (( retry++ < 10 ))
-    then
+while ! curl -sS --fail-with-body localhost:6333/collections | jq -e '.result.collections | index({"name": "low-ram"})' &>/dev/null; do
+    if ((retry++ < 10)); then
         sleep 1
     else
         echo "Collection failed to load in ~10 seconds" >&2
@@ -57,8 +77,7 @@ done
 # Check that there's a "dummy" shard log message in service logs
 declare DUMMY_SHARD_MSG='initializing "dummy" shard'
 
-if ! docker logs "$container" 2>&1 | grep "$DUMMY_SHARD_MSG"
-then
+if ! docker logs "$container" 2>&1 | grep "$DUMMY_SHARD_MSG"; then
     echo "'$DUMMY_SHARD_MSG' log message not found in $container container logs" >&2
     exit 3
 fi
@@ -66,18 +85,17 @@ fi
 # Check that there's a "low RAM" log message in service logs
 declare LOW_RAM_MSG='segment load aborted to prevent OOM'
 
-if ! docker logs "$container" 2>&1 | grep "$LOW_RAM_MSG"
-then
+if ! docker logs "$container" 2>&1 | grep "$LOW_RAM_MSG"; then
     echo "'$LOW_RAM_MSG' log message not found in $container container logs" >&2
     exit 4
 fi
 
 # Check that `low-ram` collection initialized as a "dummy" shard
 # (e.g., collection info request returns HTTP status 500)
-declare status; status="$(curl -sS localhost:6333/collections/low-ram -w ''%{http_code}'' -o /dev/null)"
+declare status
+status="$(curl -sS localhost:6333/collections/low-ram -w ''%{http_code}'' -o /dev/null)"
 
-if (( status != 500 ))
-then
+if ((status != 500)); then
     echo "Collection info request returned an unexpected HTTP status: expected 500, but received $STATUS" >&2
     exit 5
 fi
